fix delete record issue

+++ src/Form.php
@@ -208,6 +208,9 @@ class Form
         $ids = explode(',', $id);
 
         foreach ($ids as $id) {
+            if (empty($id)) {
+                continue;
+            }
             $this->deleteFilesAndImages($id);
             $this->model->find($id)->delete();
         }
